+++ src/containers/SiteLibrary/SiteLibraryRouted.js
@@ -1,26 +1,25 @@
-import React                   from 'react';
-import Typography              from '@material-ui/core/Typography';
-import { Switch, Route }       from 'react-router-dom'
-import Box                     from '@material-ui/core/Box';
-import List                    from '@material-ui/core/List';
-import ListSubheader           from '@material-ui/core/ListSubheader';
-import Grid                    from '@material-ui/core/Grid';
-import Menu                    from '@material-ui/core/Menu';
-import MenuItem                from '@material-ui/core/MenuItem';
-import IconButton              from '@material-ui/core/IconButton';
-import MoreVertIcon            from '@material-ui/icons/MoreVert';
-import RemoteSiteDialog        from './dialogs/RemoteSiteDialog';
-import ImportSiteDialog        from './dialogs/ImportSiteDialog';
-import NewSiteDialog           from './dialogs/NewSiteDialog';
-import EditTagsDialogs         from './dialogs/EditTagsDialogs';
-import RenameDialog            from './dialogs/RenameDialog';
-import SiteListItem            from './components/SiteListItem';
-import CardItem                from './components/CardItem';
-import BlockDialog             from './../../components/BlockDialog';
-import Spinner                 from './../../components/Spinner';
-import service                 from './../../services/service';
+import React                    from 'react';
+import Typography               from '@material-ui/core/Typography';
+import { Switch, Route }        from 'react-router-dom'
+import Box                      from '@material-ui/core/Box';
+import List                     from '@material-ui/core/List';
+import ListSubheader            from '@material-ui/core/ListSubheader';
+import Grid                     from '@material-ui/core/Grid';
+import Menu                     from '@material-ui/core/Menu';
+import MenuItem                 from '@material-ui/core/MenuItem';
+import IconButton               from '@material-ui/core/IconButton';
+import MoreVertIcon             from '@material-ui/icons/MoreVert';
+import RemoteSiteDialog         from './dialogs/RemoteSiteDialog';
+import NewSlashImportSiteDialog from './dialogs/NewSlashImportSiteDialog';
+import EditTagsDialogs          from './dialogs/EditTagsDialogs';
+import RenameDialog             from './dialogs/RenameDialog';
+import SiteListItem             from './components/SiteListItem';
+import CardItem                 from './components/CardItem';
+import BlockDialog              from './../../components/BlockDialog';
+import Spinner                  from './../../components/Spinner';
+import service                  from './../../services/service';
 //import { snackMessageService } from './../../services/ui-service';
-import { withStyles }          from '@material-ui/core/styles';
+import { withStyles }           from '@material-ui/core/styles';
 
 const useStyles = theme => ({
 });
@@ -37,11 +36,9 @@ class SiteLibraryRouted extends React.Component{
       editTagsDialog: false,
       renameDialog: false,
       dialogSiteConf: {},
-      dialogImportSite: {
-        open: false
-      },
-      dialogNewSite: {
-        open: false
+      dialogNewSlashImportSite: {
+        open: false,
+        newOrImport: 'new'
       },
       currentRemoteSite: '',
       publishSiteDialog: undefined,
@@ -101,11 +98,21 @@ class SiteLibraryRouted extends React.Component{
   componentWillUpdate(nextProps, nextState) {
 
     if(this.props.newSite !== nextProps.newSite){
-      this.setState({dialogNewSite: {open: nextProps.newSite}});
+      this.setState({
+        dialogNewSlashImportSite: {
+          open: nextProps.newSite,
+          newOrImport: 'new',
+        }
+      });
     }
 
     if(this.props.importSite !== nextProps.importSite){
-      this.setState({dialogImportSite: {open: nextProps.importSite}});
+      this.setState({
+        dialogNewSlashImportSite: {
+          open: nextProps.importSite,
+          newOrImport: 'import',
+        }
+      });
     }
 
     if(this.props.quiqrUsername !== nextProps.quiqrUsername){
@@ -413,17 +420,10 @@ class SiteLibraryRouted extends React.Component{
           }}
         />
 
-        <NewSiteDialog
-          open={this.state.dialogNewSite.open}
-          onClose={()=>this.setState({dialogNewSite:{open:false}})}
-          mountSite={(siteKey)=>{
-            this.mountSiteByKey(siteKey);
-          }}
-        />
-
-        <ImportSiteDialog
-          open={this.state.dialogImportSite.open}
-          onClose={()=>this.setState({dialogImportSite:{open:false}})}
+        <NewSlashImportSiteDialog
+          open={this.state.dialogNewSlashImportSite.open}
+          onClose={()=>this.setState({dialogNewSlashImportSite:{open:false}})}
+          newOrImport={this.state.dialogNewSlashImportSite.newOrImport}
           mountSite={(siteKey)=>{
             this.mountSiteByKey(siteKey);
           }}